openwrt anpassung

+++ openvpn.sh
@@ -5,8 +5,11 @@ ADMINPASSWORD="secret"
 DNS1="8.8.8.8"
 DNS2="8.8.4.4"
 PROTOCOL=udp
-PORT=1194
+PORT=65334
 HOST=$(wget -4qO- "http://whatismyip.akamai.com/")
+HTTPGIT=https://raw.githubusercontent.com/theonemule/simple-openvpn-server/master
+ADMINUSER=admin
+EASYRSAV=3.0.1
 
 
 for i in "$@"
@@ -48,10 +51,10 @@ if [[ "$EUID" -ne 0 ]]; then
 	exit 2
 fi
 
-#if [[ ! -e /dev/net/tun ]]; then
-#	echo "The TUN device is not available. You need to enable TUN before running this script."
-#	exit 3
-#fi
+if [[ ! -e /dev/net/tun ]]; then
+	echo "The TUN device is not available. You need to enable TUN before running this script."
+	exit 3
+fi
 
 if grep -qs "CentOS release 5" "/etc/redhat-release"; then
 	echo "CentOS 5 is too old and not supported"
@@ -62,12 +65,16 @@ if [[ -e /etc/debian_version ]]; then
 	OS=debian
 	GROUPNAME=nogroup
 	RCLOCAL='/etc/rc.local'
+elif [[ -e /etc/device_info ]]; then
+	OS=openwrt
+	GROUPNAME=nogroup
+	RCLOCAL='/etc/rc.local'
 elif [[ -e /etc/centos-release || -e /etc/redhat-release ]]; then
 	OS=centos
 	GROUPNAME=nobody
 	RCLOCAL='/etc/rc.d/rc.local'
 else
-	echo "Looks like you aren't running this installer on Debian, Ubuntu or CentOS"
+	echo "Looks like you aren't running this installer on Debian, openwrt,  Ubuntu or CentOS"
 	exit 5
 fi
 
@@ -83,6 +90,9 @@ fi
 if [[ "$OS" = 'debian' ]]; then
 	apt-get update
 	apt-get install openvpn iptables openssl ca-certificates lighttpd -y
+elif [[ "$OS" = 'openwrt' ]]; then
+	opkg update
+	opkg install libopenssl ca-certificates bash openvpn-openssl lighttpd 
 else
 	# Else, the distro is CentOS
 	yum install epel-release -y
@@ -95,12 +105,12 @@ if [[ -d /etc/openvpn/easy-rsa/ ]]; then
 fi
 # Get easy-rsa
 
-wget -O ~/EasyRSA-3.0.1.tgz "https://github.com/OpenVPN/easy-rsa/releases/download/3.0.1/EasyRSA-3.0.1.tgz"
-tar xzf ~/EasyRSA-3.0.1.tgz -C ~/
-mv ~/EasyRSA-3.0.1/ /etc/openvpn/
-mv /etc/openvpn/EasyRSA-3.0.1/ /etc/openvpn/easy-rsa/
+wget -O ~/EasyRSA-$EASYRSAV.tgz "https://github.com/OpenVPN/easy-rsa/releases/download/$EASYRSAV/EasyRSA-$EASYRSAV.tgz"
+tar xzf ~/EasyRSA-$EASYRSAV.tgz -C ~/
+mv ~/EasyRSA-$EASYRSAV/ /etc/openvpn/
+mv /etc/openvpn/EasyRSA-$EASYRSAV/ /etc/openvpn/easy-rsa/
 chown -R root:root /etc/openvpn/easy-rsa/
-rm -rf ~/EasyRSA-3.0.1.tgz
+rm -rf ~/EasyRSA-$EASYRSAV.tgz
 cd /etc/openvpn/easy-rsa/
 
 # Create the PKI, set up the CA, the DH params and the server + client certificates
@@ -135,77 +145,77 @@ tls-auth ta.key 0
 topology subnet
 server 10.8.0.0 255.255.255.0
 ifconfig-pool-persist ipp.txt" > /etc/openvpn/server.conf
-#echo 'push "redirect-gateway def1 bypass-dhcp"' >> /etc/openvpn/server.conf
+echo 'push "redirect-gateway def1 bypass-dhcp"' >> /etc/openvpn/server.conf
 
 # DNS
-#echo "push \"dhcp-option DNS $DNS1\"" >> /etc/openvpn/server.conf
-#echo "push \"dhcp-option DNS $DNS2\"" >> /etc/openvpn/server.conf
+echo "push \"dhcp-option DNS $DNS1\"" >> /etc/openvpn/server.conf
+echo "push \"dhcp-option DNS $DNS2\"" >> /etc/openvpn/server.conf
 echo "keepalive 10 120
 cipher AES-256-CBC
-comp-lzo
+
 user nobody
 group $GROUPNAME
 persist-key
 persist-tun
 status openvpn-status.log
 verb 3
-
-;duplicate-cn
-client-to-client
-
 crl-verify crl.pem" >> /etc/openvpn/server.conf
 
 # Enable net.ipv4.ip_forward for the system
-sed -i '/\<net.ipv4.ip_forward\>/c\net.ipv4.ip_forward=1' /etc/sysctl.conf
-if ! grep -q "\<net.ipv4.ip_forward\>" /etc/sysctl.conf; then
-	echo 'net.ipv4.ip_forward=1' >> /etc/sysctl.conf
-fi
-
-# Avoid an unneeded reboot
-echo 1 > /proc/sys/net/ipv4/ip_forward
-if pgrep firewalld; then
-	# Using both permanent and not permanent rules to avoid a firewalld
-	# reload.
-	# We don't use --add-service=openvpn because that would only work with
-	# the default port and protocol.
-	firewall-cmd --zone=public --add-port=$PORT/$PROTOCOL
-	firewall-cmd --zone=trusted --add-source=10.8.0.0/24
-	firewall-cmd --permanent --zone=public --add-port=$PORT/$PROTOCOL
-	firewall-cmd --permanent --zone=trusted --add-source=10.8.0.0/24
-	# Set NAT for the VPN subnet
-	firewall-cmd --direct --add-rule ipv4 nat POSTROUTING 0 -s 10.8.0.0/24 -j SNAT --to $IP
-	firewall-cmd --permanent --direct --add-rule ipv4 nat POSTROUTING 0 -s 10.8.0.0/24 -j SNAT --to $IP
-else
-	# Needed to use rc.local with some systemd distros
-	if [[ "$OS" = 'debian' && ! -e $RCLOCAL ]]; then
-		echo '#!/bin/sh -e
-exit 0' > $RCLOCAL
+if [[ "$OS" != 'openwrt' ]]; then
+	sed -i '/\<net.ipv4.ip_forward\>/c\net.ipv4.ip_forward=1' /etc/sysctl.conf
+	if ! grep -q "\<net.ipv4.ip_forward\>" /etc/sysctl.conf; then
+		echo 'net.ipv4.ip_forward=1' >> /etc/sysctl.conf
 	fi
-	chmod +x $RCLOCAL
-	# Set NAT for the VPN subnet
-	iptables -t nat -A POSTROUTING -s 10.8.0.0/24 -j SNAT --to $IP
-	sed -i "1 a\iptables -t nat -A POSTROUTING -s 10.8.0.0/24 -j SNAT --to $IP" $RCLOCAL
-	if iptables -L -n | grep -qE '^(REJECT|DROP)'; then
+
+	# Avoid an unneeded reboot
+	echo 1 > /proc/sys/net/ipv4/ip_forward
+	if pgrep firewalld; then
+		# Using both permanent and not permanent rules to avoid a firewalld
+		# reload.
+		# We don't use --add-service=openvpn because that would only work with
+		# the default port and protocol.
+		firewall-cmd --zone=public --add-port=$PORT/$PROTOCOL
+		firewall-cmd --zone=trusted --add-source=10.8.0.0/24
+		firewall-cmd --permanent --zone=public --add-port=$PORT/$PROTOCOL
+		firewall-cmd --permanent --zone=trusted --add-source=10.8.0.0/24
+		# Set NAT for the VPN subnet
+		firewall-cmd --direct --add-rule ipv4 nat POSTROUTING 0 -s 10.8.0.0/24 -j SNAT --to $IP
+		firewall-cmd --permanent --direct --add-rule ipv4 nat POSTROUTING 0 -s 10.8.0.0/24 -j SNAT --to $IP
+	else
+		# Needed to use rc.local with some systemd distros
+		if [[ "$OS" = 'debian' && ! -e $RCLOCAL ]]; then
+			echo '#!/bin/sh -e
+	exit 0' > $RCLOCAL
+		fi
+		chmod +x $RCLOCAL
+		# Set NAT for the VPN subnet
+		iptables -t nat -A POSTROUTING -s 10.8.0.0/24 -j SNAT --to $IP
+		sed -i "1 a\iptables -t nat -A POSTROUTING -s 10.8.0.0/24 -j SNAT --to $IP" $RCLOCAL
+		if iptables -L -n | grep -qE '^(REJECT|DROP)'; then
 		# If iptables has at least one REJECT rule, we asume this is needed.
 		# Not the best approach but I can't think of other and this shouldn't
 		# cause problems.
-		iptables -I INPUT -p $PROTOCOL --dport $PORT -j ACCEPT
-		iptables -I FORWARD -s 10.8.0.0/24 -j ACCEPT
-		iptables -I FORWARD -m state --state RELATED,ESTABLISHED -j ACCEPT
-		sed -i "1 a\iptables -I INPUT -p $PROTOCOL --dport $PORT -j ACCEPT" $RCLOCAL
-		sed -i "1 a\iptables -I FORWARD -s 10.8.0.0/24 -j ACCEPT" $RCLOCAL
-		sed -i "1 a\iptables -I FORWARD -m state --state RELATED,ESTABLISHED -j ACCEPT" $RCLOCAL
+			iptables -I INPUT -p $PROTOCOL --dport $PORT -j ACCEPT
+			iptables -I FORWARD -s 10.8.0.0/24 -j ACCEPT
+			iptables -I FORWARD -m state --state RELATED,ESTABLISHED -j ACCEPT
+			sed -i "1 a\iptables -I INPUT -p $PROTOCOL --dport $PORT -j ACCEPT" $RCLOCAL
+			sed -i "1 a\iptables -I FORWARD -s 10.8.0.0/24 -j ACCEPT" $RCLOCAL
+			sed -i "1 a\iptables -I FORWARD -m state --state RELATED,ESTABLISHED -j ACCEPT" $RCLOCAL
+		fi
 	fi
 fi
 # If SELinux is enabled and a custom port or TCP was selected, we need this
-if hash sestatus 2>/dev/null; then
-	if sestatus | grep "Current mode" | grep -qs "enforcing"; then
-		if [[ "$PORT" != '1194' || "$PROTOCOL" = 'tcp' ]]; then
-			# semanage isn't available in CentOS 6 by default
-			if ! hash semanage 2>/dev/null; then
-				yum install policycoreutils-python -y
+if [[ "$OS" != 'openwrt' ]]; then
+	if hash sestatus 2>/dev/null; then
+		if sestatus | grep "Current mode" | grep -qs "enforcing"; then
+			if [[ "$PORT" != '1194' || "$PROTOCOL" = 'tcp' ]]; then
+				# semanage isn't available in CentOS 6 by default
+				if ! hash semanage 2>/dev/null; then
+					yum install policycoreutils-python -y
+				fi
+				semanage port -a -t openvpn_port_t -p $PROTOCOL $PORT
 			fi
-			semanage port -a -t openvpn_port_t -p $PROTOCOL $PORT
 		fi
 	fi
 fi
@@ -218,6 +228,8 @@ if [[ "$OS" = 'debian' ]]; then
 	else
 		/etc/init.d/openvpn restart
 	fi
+elif [[ "$OS" = 'openwrt' ]]; then
+	/etc/init.d/openvpn restart
 else
 	if pgrep systemd-journal; then
 		systemctl restart openvpn@server.service
@@ -257,12 +269,23 @@ mkdir /etc/openvpn/clients/
 # mkdir /etc/openvpn/clients/
 
 #Set permissions for easy-rsa and open vpn to be modified by the web user.
-chown -R www-data:www-data /etc/openvpn/easy-rsa
-chown -R www-data:www-data /etc/openvpn/clients/
+if [[ "$OS"='openwrt' ]]; then
+	chown -R http:nogroup /etc/openvpn/easy-rsa
+	chown -R http:nogroup /etc/openvpn/clients/
+else
+	chown -R www-data:www-data /etc/openvpn/easy-rsa
+	chown -R www-data:www-data /etc/openvpn/clients/
+fi
 chmod -R 755 /etc/openvpn/
 chmod -R 777 /etc/openvpn/crl.pem
 chmod g+s /etc/openvpn/clients/
 chmod g+s /etc/openvpn/easy-rsa/
+
+#Set Port 443 on 8443for uhttpd openwrt
+if [[ "$OS" = 'openwrt' ]]; then
+	sed -i 's/:443/:8443/' /etc/config/uhttpd
+	/etc/init.d/uhttpd restart
+fi
 
 #Generate a self-signed certificate for the web server
 mv /etc/lighttpd/ssl/ /etc/lighttpd/ssl.$$/
@@ -273,17 +296,32 @@ chmod 744 /etc/lighttpd/ssl/server.pem
 
 #Configure the web server with the lighttpd.conf from GitHub
 mv /etc/lighttpd/lighttpd.conf /etc/lighttpd/lighttpd.conf.$$
-wget -O /etc/lighttpd/lighttpd.conf https://raw.githubusercontent.com/theonemule/simple-openvpn-server/master/lighttpd.conf
+if [[ "$OS"='openwrt' ]]; then
+	wget -O /etc/lighttpd/lighttpd-openwrt.conf $HTTPGIT/lighttpd.conf
+else
+	wget -O /etc/lighttpd/lighttpd.conf $HTTPGIT/lighttpd.conf
+fi
 
 #install the webserver scripts
-rm /var/www/html/*
-wget -O /var/www/html/index.sh https://raw.githubusercontent.com/theonemule/simple-openvpn-server/master/index.sh
+if [[ "$OS"='openwrt' ]]; then
+	mkdir /www2
+	wget -O /www2/index.sh $HTTPGIT/index.sh
+	wget -O /www2/download.sh $HTTPGIT/download.sh
+	chown -R http:nogroup /www2
+else
+	rm /var/www/html/*
+	wget -O /var/www/html/index.sh $HTTPGIT/index.sh
 
-wget -O /var/www/html/download.sh https://raw.githubusercontent.com/theonemule/simple-openvpn-server/master/download.sh
-chown -R www-data:www-data /var/www/html/
+	wget -O /var/www/html/download.sh $HTTPGIT/download.sh
+	chown -R www-data:www-data /var/www/html/
+fi
 
 #set the password file for the WWW logon
-echo "admin:$ADMINPASSWORD" >> /etc/lighttpd/.lighttpdpassword
+echo "$ADMINUSER:$ADMINPASSWORD" >> /etc/lighttpd/.lighttpdpassword
 
 #restart the web server
-service lighttpd restart
+if [[ "$OS" = 'openwrt' ]]; then
+	/etc/init.d/lighttpd restart
+else
+	service lighttpd restart
+fi
